requirements_bkp.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.2
botocore==1.8.28
cement==2.8.2
certifi==14.5.14
colorama==0.3.7
configparser==3.5.0
decorator==4.1.2
Django==1.9.7
django-adminfiles==1.0.1
django-allauth==0.34.0
django-ckeditor==5.0.3
django-cleanup==0.3.1
django-crispy-forms==1.6.0
django-haystack==2.4.0
django-localflavor==1.0
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-pygments==0.1
django-templateaddons==0.1
django-tinymce==2.3.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
html5lib==1.0.1
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==2.6.6
markdown2==2.3.1
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
oauthlib==2.0.6
pandocfilters==1.4.2
pathlib2==2.3.0
pathspec==0.3.4
pickleshare==0.7.4
Pillow==2.8.1
prompt-toolkit==1.0.15
Pygments==2.0.2
python-dateutil==2.6.1
python-openid==2.2.5
PyYAML==3.11
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.9.1
requests-oauthlib==0.8.0
scandir==1.6
semantic-version==2.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
testpath==0.3.1
texttable==0.8.8
tornado==4.5.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.46.0
Whoosh==2.7.0
widgetsnbextension==3.1.0
win-unicode-console==0.5
